refactor order of ops
