+++ local/iatlas/cbioportal_export/clinical.py
@@ -5,12 +5,10 @@ import logging
 from pathlib import Path
 import os
 import subprocess
-import sys
 from typing import Dict, List
 
 import pandas as pd
 import synapseclient
-import synapseutils
 
 import utils
 
@@ -425,7 +423,7 @@ def convert_oncotree_codes(datahub_tools_path: str) -> pd.DataFrame:
 
 
 def get_all_non_na_columns(input_df: pd.DataFrame) -> List[str]:
-    """ Gets all the columns in input data without all (100%) NAs
+    """Gets all the columns in input data without all (100%) NAs
     Args:
         input_df (pd.DataFrame): input data
 
@@ -466,7 +464,7 @@ def add_clinical_header(
     sample_df_subset = input_dfs["sample"][
         input_dfs["sample"]["Dataset"] == dataset_name
     ]
-    
+
     # get the columns without 100% NAs
     patient_subset_cols = get_all_non_na_columns(input_df=patient_df_subset)
     sample_subset_cols = get_all_non_na_columns(input_df=sample_df_subset)
@@ -677,99 +675,6 @@ def write_case_lists_all_and_sequenced(
     subprocess.run(cmd, shell=True, executable="/bin/bash")
 
 
-def save_to_synapse(
-    dataset_name: str,
-    datahub_tools_path: str,
-    output_folder_synid: str,
-    version_comment: str = None,
-) -> None:
-    """Saves the dataset's clinical file, case lists
-        and meta files to its synapse respective folders
-
-    Args:
-        dataset_name (str): name of the iatlas dataset to save to
-            synapse
-        datahub_tools_path (str): Path to the datahub tools repo
-        output_folder_synid (str): Synapse id of the output folder
-        version_comment (str): Version comment for this iteration of files on synapse. Optional.
-            Defaults to None.
-    """
-    # TODO: Make into argument
-    dataset_dir = os.path.join(datahub_tools_path, "add-clinical-header", dataset_name)
-    # see if dataset_folder exists
-    dataset_folder_exists = False
-    for _, directory_names, _ in synapseutils.walk(syn=syn, synId=output_folder_synid):
-        directories = directory_names  # top level directories
-        break
-
-    for dataset_folder in directories:
-        if dataset_name == dataset_folder[0]:
-            dataset_folder_exists = True
-            dataset_folder_id = dataset_folder[1]
-            break
-
-    if not dataset_folder_exists:
-        new_dataset_folder = synapseclient.Folder(
-            dataset_name, parent=output_folder_synid
-        )
-        dataset_folder_id = syn.store(new_dataset_folder).id
-
-    # store clinical patient file
-    syn.store(
-        synapseclient.File(
-            f"{dataset_dir}/data_clinical_patient.txt",
-            name="data_clinical_patient.txt",
-            parent=dataset_folder_id,
-            version_comment=version_comment,
-        )
-    )
-    # store clinical sample file
-    syn.store(
-        synapseclient.File(
-            f"{dataset_dir}/data_clinical_sample.txt",
-            name="data_clinical_sample.txt",
-            parent=dataset_folder_id,
-            version_comment=version_comment,
-        )
-    )
-    # store meta* files
-    syn.store(
-        synapseclient.File(
-            f"{dataset_dir}/meta_clinical_patient.txt",
-            parent=dataset_folder_id,
-            version_comment=version_comment,
-        )
-    )
-    syn.store(
-        synapseclient.File(
-            f"{dataset_dir}/meta_clinical_sample.txt",
-            parent=dataset_folder_id,
-            version_comment=version_comment,
-        )
-    )
-    syn.store(
-        synapseclient.File(
-            f"{dataset_dir}/meta_study.txt",
-            parent=dataset_folder_id,
-            version_comment=version_comment,
-        )
-    )
-    case_list_files = os.listdir(os.path.join(dataset_dir, "case_lists"))
-    case_list_folder = synapseclient.Folder("case_lists", parent=dataset_folder_id)
-    try:
-        case_list_folder_id = syn.store(case_list_folder).id
-    except:
-        sys.exit(-1)
-    for file in case_list_files:
-        syn.store(
-            synapseclient.File(
-                f"{dataset_dir}/case_lists/{file}",
-                parent=case_list_folder_id,
-                version_comment=version_comment,
-            )
-        )
-
-
 def validate_export_files(
     input_df_synid: str, dataset_name: str, datahub_tools_path: str, **kwargs
 ) -> None:
@@ -795,7 +700,7 @@ def validate_export_files(
     dataset_dir = utils.get_local_dataset_output_folder_path(
         dataset_name, datahub_tools_path
     )
-    
+
     for file in REQUIRED_OUTPUT_FILES:
         if file.startswith("cases"):
             required_file_path = f"{dataset_dir}/case_lists/{file}"
@@ -837,45 +742,15 @@ def validate_export_files(
 
     if output_patient_df.PATIENT_ID.isna().any():
         logger.error("There are missing PATIENT_ID values.")
-    
+
     # check that there are no all NA columns
     if output_patient_df.isna().all().any():
         logger.error("There are patient columns with ALL NAs.")
-        
+
     if output_samples_df.isna().all().any():
         logger.error("There are sample columns with ALL NAs.")
-        
+
     print("\n\n")
-
-
-def run_cbioportal_validator(
-    dataset_name: str, cbioportal_path: str, datahub_tools_path: str
-) -> None:
-    """Runs the cbioportal validation script to check the
-        input clinical, metadata files and saves the output
-
-    Args:
-        dataset_name (str): name of the dataset
-        cbioportal_path (str): Path to cbioportal repo containing validator script
-        datahub_tools_path (str): path to the datahub tools repo containing
-            the locally saved clinical files
-    """
-    cmd = f"""
-    python3 {cbioportal_path}/core/src/main/scripts/importer/validateData.py \
-        -s "{datahub_tools_path}/add-clinical-header/{dataset_name}" \
-            --no_portal_checks \
-            --strict_maf_checks
-    """
-    validated = f"{datahub_tools_path}/add-clinical-header/{dataset_name}/cbioportal_validator_output.txt"
-    with open(f"{validated}", "w") as outfile:
-        subprocess.run(
-            cmd,
-            shell=True,
-            executable="/bin/bash",
-            stdout=outfile,
-            stderr=subprocess.STDOUT,
-        )
-    print(f"cbioportal validator results saved to: {validated}")
 
 
 def main():
@@ -907,37 +782,15 @@ def main():
         help="Synapse id for the study_sample_name (paper ids) to lens id mapping file",
     )
     parser.add_argument(
-        "--output_folder_synid",
-        type=str,
-        help="Synapse id for output folder to store the export files",
-    )
-    parser.add_argument(
         "--datahub_tools_path",
         type=str,
         help="Path to datahub-study-curation-tools repo",
-    )
-    parser.add_argument(
-        "--cbioportal_path",
-        type=str,
-        help="Path to cbioportal repo",
-    )
-    parser.add_argument(
-        "--dry_run",
-        action="store_true",
-        default=False,
-        help="Whether to run without saving to Synapse",
     )
     parser.add_argument(
         "--clear_workspace",
         action="store_true",
         default=False,
         help="Whether to clear local directory of files or not",
-    )
-    parser.add_argument(
-        "--version_comment",
-        default=None,
-        type=str,
-        help="Version comment for the files on Synapse. Optional. Defaults to None.",
     )
 
     args = parser.parse_args()
@@ -998,18 +851,6 @@ def main():
             datahub_tools_path=args.datahub_tools_path,
             logger=dataset_logger,
         )
-        run_cbioportal_validator(
-            dataset_name=dataset,
-            cbioportal_path=args.cbioportal_path,
-            datahub_tools_path=args.datahub_tools_path,
-        )
-        if not args.dry_run:
-            save_to_synapse(
-                dataset_name=dataset,
-                datahub_tools_path=args.datahub_tools_path,
-                output_folder_synid=args.output_folder_synid,
-                version_comment=args.version_comment,
-            )
 
 
 if __name__ == "__main__":
