+++ local/iatlas/cbioportal_export/load.py
@@ -1,0 +1,177 @@
+import argparse
+import os
+import subprocess
+import sys
+
+import synapseclient
+import synapseutils
+
+
+my_agent = "iatlas-cbioportal/0.0.0"
+syn = synapseclient.Synapse(user_agent=my_agent).login()
+
+
+REQUIRED_OUTPUT_FILES = [
+    "data_clinical_patient.txt",
+    "data_clinical_sample.txt",
+    "meta_clinical_patient.txt",
+    "meta_clinical_sample.txt",
+    "data_mutations.txt",
+    "meta_mutations.txt",
+    "data_gene_signatures.txt",
+    "meta_gene_signatures.txt",
+    "data_rna_seq_mrna.txt",
+    "meta_rna_seq_mrna.txt",
+    "data_neoantigen_priority_scores.txt",
+    "meta_eoantigen_priority_scores.txt",
+    "cbioportal_validator_output.txt"
+]
+
+
+def run_cbioportal_validator(
+    dataset_name: str, cbioportal_path: str, datahub_tools_path: str
+) -> None:
+    """Runs the cbioportal validation script to check the
+        input clinical, metadata files and saves the output
+
+    Args:
+        dataset_name (str): name of the dataset
+        cbioportal_path (str): Path to cbioportal repo containing validator script
+        datahub_tools_path (str): path to the datahub tools repo containing
+            the locally saved clinical files
+    """
+    cmd = f"""
+    python3 {cbioportal_path}/core/src/main/scripts/importer/validateData.py \
+        -s "{datahub_tools_path}/add-clinical-header/{dataset_name}" \
+            --no_portal_checks \
+            --strict_maf_checks
+    """
+    validated = f"{datahub_tools_path}/add-clinical-header/{dataset_name}/cbioportal_validator_output.txt"
+    with open(f"{validated}", "w") as outfile:
+        subprocess.run(
+            cmd,
+            shell=True,
+            executable="/bin/bash",
+            stdout=outfile,
+            stderr=subprocess.STDOUT,
+        )
+    print(f"cbioportal validator results saved to: {validated}")
+
+
+def save_to_synapse(
+    dataset_name: str,
+    datahub_tools_path: str,
+    output_folder_synid: str,
+    version_comment: str = None,
+) -> None:
+    """Saves the dataset's clinical file, case lists
+        and meta files to its synapse respective folders
+
+    Args:
+        dataset_name (str): name of the iatlas dataset to save to
+            synapse
+        datahub_tools_path (str): Path to the datahub tools repo
+        output_folder_synid (str): Synapse id of the output folder
+        version_comment (str): Version comment for this iteration of files on synapse. Optional.
+            Defaults to None.
+    """
+    # TODO: Make into argument
+    dataset_dir = os.path.join(datahub_tools_path, "add-clinical-header", dataset_name)
+    # see if dataset_folder exists
+    dataset_folder_exists = False
+    for _, directory_names, _ in synapseutils.walk(syn=syn, synId=output_folder_synid):
+        directories = directory_names  # top level directories
+        break
+
+    for dataset_folder in directories:
+        if dataset_name == dataset_folder[0]:
+            dataset_folder_exists = True
+            dataset_folder_id = dataset_folder[1]
+            break
+
+    if not dataset_folder_exists:
+        new_dataset_folder = synapseclient.Folder(
+            dataset_name, parent=output_folder_synid
+        )
+        dataset_folder_id = syn.store(new_dataset_folder).id
+
+    # store required files
+    for file in REQUIRED_OUTPUT_FILES:
+        syn.store(
+            synapseclient.File(
+                f"{dataset_dir}/{file}",
+                name=file,
+                parent=dataset_folder_id,
+                version_comment=version_comment
+            )
+        )
+    
+    # store case lists
+    case_list_files = os.listdir(os.path.join(dataset_dir, "case_lists"))
+    case_list_folder = synapseclient.Folder("case_lists", parent=dataset_folder_id)
+    try:
+        case_list_folder_id = syn.store(case_list_folder).id
+    except:
+        sys.exit(-1)
+    for file in case_list_files:
+        syn.store(
+            synapseclient.File(
+                f"{dataset_dir}/case_lists/{file}",
+                parent=case_list_folder_id,
+                version_comment=version_comment,
+            )
+        )
+    
+
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "--dataset",
+        type=str,
+        help="Name of dataset to run processing for",
+    )
+    parser.add_argument(
+        "--output_folder_synid",
+        type=str,
+        help="Synapse id for output folder to store the export files",
+    )
+    parser.add_argument(
+        "--datahub_tools_path",
+        type=str,
+        help="Path to datahub-study-curation-tools repo",
+    )
+    parser.add_argument(
+        "--cbioportal_path",
+        type=str,
+        help="Path to cbioportal repo",
+    )
+    parser.add_argument(
+        "--dry_run",
+        action="store_true",
+        default=False,
+        help="Whether to run without saving to Synapse",
+    )
+    parser.add_argument(
+        "--version_comment",
+        default=None,
+        type=str,
+        help="Version comment for the files on Synapse. Optional. Defaults to None.",
+    )
+    args = parser.parse_args()
+    
+    run_cbioportal_validator(
+        dataset_name=args.dataset,
+        cbioportal_path=args.cbioportal_path,
+        datahub_tools_path=args.datahub_tools_path,
+    )
+    if not args.dry_run:
+        save_to_synapse(
+            dataset_name=args.dataset,
+            datahub_tools_path=args.datahub_tools_path,
+            output_folder_synid=args.output_folder_synid,
+            version_comment=args.version_comment,
+        )
+
+
+if __name__ == "__main__":
+    main()
